exclusão da tela confirmação cadastro, componentização do input de texto, e tela de inicio, string.xml da tela de inico e cadastro personal
