+++ .gradle/config.properties
@@ -1,2 +1,2 @@
-#Mon Sep 23 18:52:09 BRT 2024
-java.home=C\:\\Users\\Cliente\\AppData\\Local\\Programs\\Android Studio\\jbr
+#Fri Oct 11 16:36:33 BRT 2024
+java.home=C\:\\Users\\alo cell\\AppData\\Local\\Programs\\Android Studio\\jbr
