+++ app/src/main/java/com/example/vitalisapp/Activity/Componentes.kt
@@ -1,5 +1,6 @@
 package com.example.vitalisapp.Activity
 
+import android.content.Intent
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -18,10 +19,14 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentWidth
+import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -31,9 +36,12 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
@@ -194,6 +202,20 @@ fun MenuPersonal(navController: NavController) {
     }
 }
 
+@Composable
+fun InputText(
+    value: String,
+    onValueChange: (String) -> Unit,
+    label: String,
+    modifier: Modifier = Modifier
+) {
+    OutlinedTextField(
+        value = value,
+        onValueChange = onValueChange,
+        label = { Text(text = label, fontFamily = MavenPro, color = Color.White) },
+        modifier = modifier.fillMaxWidth()
+    )
+}
 
 @Composable
 fun ImageCard(modifier: Modifier = Modifier, imageRes: Int, date: String) {
@@ -329,6 +351,150 @@ fun CardConversa(
                         fontFamily = MavenPro,
                         color = Color.Gray,
                         fontSize = 5.sp
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun Atividade() {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .verticalScroll(rememberScrollState())
+            .padding(16.dp)
+    ) {
+        Text(
+            text = stringResource(R.string.atividade),
+            fontSize = 20.sp,
+            fontFamily = MavenPro,
+            fontWeight = FontWeight.Bold,
+            color = Color.Black,
+            modifier = Modifier
+                .padding(bottom = 15.dp)
+                .align(Alignment.CenterHorizontally)
+        )
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .background(Color(18, 18, 19), shape = RoundedCornerShape(24.dp)),
+            horizontalArrangement = Arrangement.spacedBy(5.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Porcentagem(
+                icone = R.mipmap.comidappreto,
+                valor = "--/10",
+                titulo = "Refeições"
+            )
+            Porcentagem(
+                icone = R.mipmap.exerciciopreto,
+                valor = "01/10",
+                titulo = "Exercícios"
+            )
+            Porcentagem(
+                icone = R.mipmap.calendario,
+                valor = "01/03",
+                titulo = "Meta Semanal"
+            )
+        }
+        Spacer(modifier = Modifier.height(16.dp))
+        CardAtividades(titulo="Título")
+        Spacer(modifier = Modifier.weight(1f))
+    }
+}
+
+@Composable
+fun Porcentagem(icone: Int, valor: String, titulo: String) {
+    Column(
+        modifier = Modifier
+            .padding(8.dp)
+            .size(90.dp)
+            .background(Color.White, shape = RoundedCornerShape(16.dp))
+            .shadow(elevation = 2.dp, shape = RoundedCornerShape(16.dp)),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center
+    ) {
+        Image(
+            painter = painterResource(id = icone),
+            contentDescription = null,
+            modifier = Modifier
+                .size(24.dp)
+                .align(Alignment.CenterHorizontally)
+        )
+        Spacer(modifier = Modifier.height(4.dp))
+        Text(
+            text = valor,
+            fontFamily = MavenPro,
+            fontSize = 18.sp,
+            fontWeight = FontWeight.SemiBold,
+            color = Color.Black
+        )
+        Spacer(modifier = Modifier.height(4.dp))
+        Text(
+            text = titulo,
+            fontSize = 12.sp,
+            fontFamily = MavenPro,
+            fontWeight = FontWeight.Medium,
+            color = Color.Black
+        )
+    }
+}
+
+@Composable
+fun CardAtividades( titulo: String) {
+    val contexto = LocalContext.current
+    Column {
+        Card(
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(70.dp)
+                .shadow(elevation = 5.dp, shape = RoundedCornerShape(16.dp)),
+            shape = RoundedCornerShape(16.dp)
+        ) {
+            Row(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(start = 16.dp),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Row(
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.spacedBy(10.dp)
+                ) {
+                    Image(
+                        painter = painterResource(id = R.mipmap.tortadefrango),
+                        contentDescription = "Exercicio",
+                        modifier = Modifier
+                            .size(60.dp)
+                            .shadow(elevation = 5.dp, shape = RoundedCornerShape(16.dp))
+                    )
+                    Text(
+                        text = stringResource(R.string.exercicio),
+                        fontSize = 18.sp,
+                        fontFamily = MavenPro,
+                        fontWeight = FontWeight.SemiBold,
+                        color = Color(72, 183, 90)
+                    )
+                    Text(
+                        text = titulo,
+                        fontFamily = MavenPro,
+                        fontSize = 18.sp,
+                        fontWeight = FontWeight.SemiBold,
+                        color = Color.Black
+                    )
+                }
+                IconButton(
+                    onClick = {val detalheExercicio = Intent(contexto, DetalheExercicio::class.java)
+                        contexto.startActivity(detalheExercicio)},
+                    modifier = Modifier.padding(end = 16.dp)
+                ) {
+                    Image(
+                        painter = painterResource(id = R.mipmap.seta),
+                        contentDescription = "seta",
+                        modifier = Modifier.size(34.dp)
                     )
                 }
             }
