+++ app/src/main/java/com/example/vitalisapp/Activity/EncontreAcademia.kt
@@ -85,136 +85,12 @@ fun BuscaAcademia(name: String, navController: NavHostController, modifier: Modi
 
         item {
             Text(
-                text = "Encontre uma academia",
+                text = "Encontre um personal",
                 fontSize = 32.sp,
                 fontWeight = FontWeight.SemiBold,
                 color = Color(72, 183, 90),
-                modifier = Modifier.padding(bottom = 16.dp)
+                modifier = Modifier.padding(vertical = 16.dp)
             )
-        }
-
-        item {
-            Row(modifier = Modifier.padding(bottom = 16.dp).fillMaxWidth()) {
-                Column(modifier = Modifier.weight(1f).padding(end = 8.dp)) {
-                    Text(
-                        text = "CEP",
-                        color = Color(0xFF1A1A1A),
-                        fontSize = 12.sp,
-                        fontWeight = FontWeight.Bold,
-                        letterSpacing = 0.6.sp,
-                        modifier = Modifier.padding(bottom = 8.dp)
-                    )
-                    OutlinedTextField(
-                        value = cep,
-                        onValueChange = { cep = it },
-                        shape = RoundedCornerShape(30.dp),
-                        modifier = Modifier.fillMaxWidth().height(40.dp),
-                        placeholder = { Text(text = "CEP", fontSize = 10.sp, color = Color.Gray) }
-                    )
-                }
-
-                Column(modifier = Modifier.weight(1f).padding(start = 8.dp)) {
-                    Text(
-                        text = "Bairro",
-                        color = Color(0xFF1A1A1A),
-                        fontSize = 12.sp,
-                        fontWeight = FontWeight.Bold,
-                        letterSpacing = 0.6.sp,
-                        modifier = Modifier.padding(bottom = 8.dp)
-                    )
-                    OutlinedTextField(
-                        value = bairro,
-                        onValueChange = { bairro = it },
-                        shape = RoundedCornerShape(30.dp),
-                        modifier = Modifier.fillMaxWidth().height(40.dp),
-                        placeholder = { Text(text = "Bairro", fontSize = 10.sp, color = Color.Gray) }
-                    )
-                }
-            }
-        }
-
-        item {
-            Row(modifier = Modifier.padding(bottom = 16.dp).fillMaxWidth()) {
-                Column(modifier = Modifier.weight(1f).padding(end = 8.dp)) {
-                    Text(
-                        text = "Rua",
-                        color = Color(0xFF1A1A1A),
-                        fontSize = 12.sp,
-                        fontWeight = FontWeight.Bold,
-                        letterSpacing = 0.6.sp,
-                        modifier = Modifier.padding(bottom = 8.dp)
-                    )
-                    OutlinedTextField(
-                        value = rua,
-                        onValueChange = { rua = it },
-                        shape = RoundedCornerShape(30.dp),
-                        modifier = Modifier.fillMaxWidth().height(40.dp),
-                        placeholder = { Text(text = "Rua", fontSize = 10.sp, color = Color.Gray) }
-                    )
-                }
-
-                Column(modifier = Modifier.weight(1f).padding(start = 8.dp)) {
-                    Text(
-                        text = "Cidade",
-                        color = Color(0xFF1A1A1A),
-                        fontSize = 12.sp,
-                        fontWeight = FontWeight.Bold,
-                        letterSpacing = 0.6.sp,
-                        modifier = Modifier.padding(bottom = 8.dp)
-                    )
-                    OutlinedTextField(
-                        value = cidade,
-                        onValueChange = { cidade = it },
-                        shape = RoundedCornerShape(30.dp),
-                        modifier = Modifier.fillMaxWidth().height(40.dp),
-                        placeholder = { Text(text = "Cidade", fontSize = 10.sp, color = Color.Gray) }
-                    )
-                }
-            }
-        }
-
-        item {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(bottom = 10.dp),
-                horizontalArrangement = Arrangement.End
-            ) {
-                Button(
-                    onClick = { },
-                    modifier = Modifier
-                        .width(100.dp),
-                    colors = ButtonDefaults.buttonColors(containerColor = Color(72, 183, 90))
-                ) {
-                    Text("Buscar", color = Color.White)
-                }
-            }
-        }
-
-        item {
-            Image(
-                painter = painterResource(id = R.mipmap.mapa),
-                contentDescription = "Mapa",
-                contentScale = ContentScale.Crop,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .aspectRatio(1.5f)
-                    .padding(8.dp)
-                    .clip(RoundedCornerShape(16.dp))
-            )
-        }
-
-        item {
-            AcademiaCard(
-                nome = "Academia Gaviões",
-                avaliacao = "4.7",
-                horario = "Seg a Sex de 06h30 às 23h00; Sáb., Dom. e Feriados de 09h30 às 18h00",
-                endereco = "Rua dos Artistas, 416",
-                tempoCaminhada = "10 min de caminhada"
-            )
-        }
-
-        item {
             PersonalCard(
                 nome = "Thiago",
                 especialidade = "Emagrecimento",
@@ -222,83 +98,6 @@ fun BuscaAcademia(name: String, navController: NavHostController, modifier: Modi
             )
         }
     }
-}
-
-@Composable
-fun AcademiaCard(
-    nome: String,
-    avaliacao: String,
-    horario: String,
-    endereco: String,
-    tempoCaminhada: String
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(bottom = 16.dp)
-            .shadow(elevation = 4.dp, shape = RoundedCornerShape(16.dp)),
-        shape = RoundedCornerShape(16.dp)
-    ) {
-        Box(
-            modifier = Modifier
-                .background(Color.White)
-                .padding(16.dp)
-        ) {
-            Column(modifier = Modifier.padding(16.dp)) {
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        text = nome,
-                        fontSize = 20.sp,
-                        color = Color(72, 183, 90),
-                        fontWeight = FontWeight.Bold,
-                        modifier = Modifier.weight(1f)
-                    )
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Image(
-                            painter = painterResource(id = R.mipmap.estrela),
-                            contentDescription = "Estrela",
-                            modifier = Modifier.size(16.dp)
-                        )
-                        Spacer(modifier = Modifier.width(4.dp))
-                        Text(
-                            text = avaliacao,
-                            fontSize = 16.sp,
-                            color = Color(255, 255, 0),
-                            fontWeight = FontWeight.Bold
-                        )
-                    }
-                }
-                Spacer(modifier = Modifier.height(8.dp))
-                Text(
-                    text = horario,
-                    fontSize = 14.sp,
-                    color = Color.Black
-                )
-                Text(
-                    text = endereco,
-                    fontSize = 14.sp,
-                    color = Color.Black
-                )
-                Text(
-                    text = tempoCaminhada,
-                    fontSize = 14.sp,
-                    color = Color.Black
-                )
-            }
-        }
-    }
-    Text(
-        text = "Encontre um personal",
-        fontSize = 32.sp,
-        fontWeight = FontWeight.SemiBold,
-        color = Color(72, 183, 90),
-        modifier = Modifier.padding(vertical = 16.dp)
-    )
 }
 
 @Composable
