+++ app/src/main/java/com/example/vitalisapp/Activity/ListaExercicio.kt
@@ -22,12 +22,14 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
+import com.example.vitalisapp.R
 import com.example.vitalisapp.ui.theme.MavenPro
 import com.example.vitalisapp.ui.theme.VitalisAppTheme
 
@@ -59,7 +61,7 @@ fun GaleriaExercicio(name: String, navController: NavHostController, modifier: M
             Menu(navController)
 
             Text(
-                text = "Exercícios - 13/05/2024",
+                text = stringResource(R.string.exercicio),
                 fontFamily = MavenPro,
                 fontSize = 25.sp,
                 fontWeight = FontWeight.Bold,
