+++ app/src/main/res/values/strings.xml
@@ -22,4 +22,16 @@
     <string name="title_activity_chat_usuario">ChatUsuario</string>
     <string name="title_activity_chat_personal">ChatPersonal</string>
     <string name="title_activity_home_personal">HomePersonal</string>
+
+    <string name="cabecalho_intrutor">Bem-vindo à nossa plataforma! Cadastre-se para acessar recursos e trabalhar conosco.</string>
+    <string name="instrutor">É um instrutor?</string>
+    <string name="cadastro_instrutor">Cadastro Instrutor</string>
+    <string name="sub_instrutor1">Quer acessar nossa aplicação? Vamos realizar seu cadastro! </string>
+    <string name="sub_instrutor2">Insira algumas informações sobre você para fazermos o cadastro de sua conta! </string>
+
+
+    <string name="Bem_vindo">Bem vindo(a), %s</string>
+    <string name="atividade">Atividades de hoje</string>
+    <string name="exercicio">Exercícios</string>
+
 </resources>
